Some small changes... :sweat_smile:
* Move all iris code into iris_bot.
* Add __main__.py in order to run iris_bot as module.
* Split commands.py into commands package
* Make commands package automatically import command functions from sub-modules

The automatic command loader expects the command to share the name of the module
(e.g. `random.py` should have a function called `random`). Alternatively, if the
module intends to export other methods, it can define `__all__`. For example,
in practice, the random module only has `roll` and `flip` functions. Therefore,
it defines `__all__ = ["flip", "roll"]`. The automatic command loader then loads
these functions.
