+++ iris_bot/__main__.py
@@ -1,0 +1,1 @@
+from . import bot
