+++ iris_bot/utils/gym.py
@@ -1,0 +1,5 @@
+from ..models import db_session, User, GymToken, GymNotification
+
+
+def check_for_classes():
+    return ''
