--- utils/gym.py
@@ -1,7 +1,0 @@
-import os
-import discord
-import json
-from models import db_session, User, GymToken, GymNotification
-
-def check_for_classes():
-    return ''
